Add Args and Kwargs Project
In Python, you could use *args and **kwargs for optional lists of arguments but I was trying to figure out how to do it in VB.NET so I made this project for testing purposes. I found that it is as simple as using ParamArray. Though, I have yet to code kwargs into the program.