Add Custom Console Functions Project
Coded some custom console functions like input and print so i can change text colors and have a more similar syntax to Python. Especially the input function that allows me to enter a string as argument to output without the need of an extra line to print out the message, just like Python.